initial pass at validation and fixing specs
