+++ Guardfile
@@ -1,37 +1,29 @@
 # frozen_string_literal: true
 
-# A sample Guardfile
-# More info at https://github.com/guard/guard#readme
-
-## Uncomment and set this to only include directories you want to watch
-# directories %w(app lib config test spec features) \
-#  .select{|d| Dir.exists?(d) ? d : UI.warning("Directory #{d} does not exist")}
-
-## Note: if you are using the `directories` clause above and you are not
-## watching the project directory ('.'), then you will want to move
-## the Guardfile to a watched dir and symlink it back, e.g.
-#
-#  $ mkdir config
-#  $ mv Guardfile config/
-#  $ ln -s config/Guardfile .
-#
-# and, you'll have to watch "config/Guardfile" instead of "Guardfile"
-
-# Note: The cmd option is now required due to the increasing number of ways
-#       rspec may be run, below are examples of the most common uses.
-#  * bundler: 'bundle exec rspec'
-#  * bundler binstubs: 'bin/rspec'
-#  * spring: 'bin/rspec' (This will use spring if running and you have
-#                          installed the spring binstubs per the docs)
-#  * zeus: 'zeus rspec' (requires the server to be started separately)
-#  * 'just' rspec: 'rspec'
-
 guard :rspec, cmd: 'NOCOVERAGE=y bin/rspec' do
-  watch(%r{spec/.*/}) { 'spec' }
-  watch(%r{app/.*/}) { 'spec' }
+  require 'guard/rspec/dsl'
+  dsl = Guard::RSpec::Dsl.new(self)
+
+  watch(dsl.rspec.spec_helper) { dsl.rspec.spec_dir }
+  watch(dsl.rspec.spec_support) { dsl.rspec.spec_dir }
+  watch(dsl.rspec.spec_files)
+
+  dsl.watch_spec_files_for(dsl.ruby.lib_files)
+  dsl.watch_spec_files_for(dsl.rails.app_files)
+
+  watch(dsl.rails.controllers) do |m|
+    [
+      dsl.rspec.spec.call("routing/#{m[1]}_routing"),
+      dsl.rspec.spec.call("controllers/#{m[1]}_controller")
+    ]
+  end
+
+  watch(dsl.rails.spec_helper) { rspec.spec_dir }
+  watch(dsl.rails.routes) { "#{rspec.spec_dir}/routing" }
+  watch(dsl.rails.app_controller) { "#{rspec.spec_dir}/controllers" }
 end
 
 guard :rubocop, cli: %w[--auto-correct] do
-  watch(/.+\.rb$/)
-  watch(%r{(?:.+/)?\.rubocop\.yml$}) { |m| File.dirname(m[0]) }
+  watch(%r{/.+\.rb$/})
+  watch(%r{/(?:.+\/)?\.rubocop\.yml$/}) { |m| File.dirname(m[0]) }
 end
