+++ lib/facilities/configuration.rb
@@ -2,6 +2,7 @@
 
 require 'common/client/configuration/rest'
 require 'common/client/middleware/response/raise_error'
+require 'common/client/middleware/response/facility_validator'
 require 'common/client/middleware/response/facility_parser'
 require 'typhoeus'
 require 'typhoeus/adapters/faraday'
@@ -28,6 +29,7 @@ module Facilities
 
         conn.response :raise_error, error_prefix: service_name
         conn.response :facility_parser
+        conn.response :facility_validator
 
         conn.adapter Faraday.default_adapter
       end
