+++ spec/lib/facilities/nca_facility_spec.rb
@@ -3,6 +3,9 @@
 require 'rails_helper'
 module Facilities
   RSpec.describe NCAFacility do
+    before(:each) { Facilities::FacilityMapping.validate_on_load = false }
+    after(:each) { Facilities::FacilityMapping.validate_on_load = true }
+
     it 'should be an NCAFacility object' do
       expect(described_class.new).to be_an(NCAFacility)
     end
