+++ spec/lib/facilities/vc_facility_spec.rb
@@ -3,6 +3,9 @@
 require 'rails_helper'
 module Facilities
   RSpec.describe VCFacility do
+    before(:each) { Facilities::FacilityMapping.validate_on_load = false }
+    after(:each) { Facilities::FacilityMapping.validate_on_load = true }
+
     it 'should be a VCFacility object' do
       expect(described_class.new).to be_a(VCFacility)
     end
