+++ spec/lib/facilities/vha_facility_spec.rb
@@ -3,6 +3,8 @@
 require 'rails_helper'
 module Facilities
   RSpec.describe VHAFacility do
+    before(:each) { Facilities::FacilityMapping.validate_on_load = false }
+    after(:each) { Facilities::FacilityMapping.validate_on_load = true }
     it 'should be a VHAFacility object' do
       expect(described_class.new).to be_a(VHAFacility)
     end
